Rename index.js for clarity
